Rewrite of shapes.exe, GButton.h .cpp to get GButton working properly

+++ inc/GTextBox.h
@@ -13,7 +13,8 @@ class GTextBox : public GComponent
     int width = 10;
     int height = 10;
     static const int TEXTSIZE = 20;
-    static const size_t CONTENT_MAX_SIZE = 3;
+    //static const size_t CONTENT_MAX_SIZE = 3;
+    static const int CONTENT_MAX_SIZE = 3;
     char content[CONTENT_MAX_SIZE + 1] = "\0";
     int contentCurrentSize = 0;
